Minor improvements
Generating keys did not work too nicely; the `generate_keys.sh` script
was changed to print the current command and to not require key
encryption. Furthermore, the keystore passphrase was changed to the
de dacto default 'changeit'.

`.gitignore` file was added to ensure the repository stays clean.

`TLSConfig.java` was changed to work with the `config/*` files
out of the box.

README is now in Markdown format.